// src/types/theme.ts
export type Theme = "aqua" | "rainbow" | "accessibility";

export interface ThemeConfig {
  bg: string;
  card: string;
  cardDark: string;
  accent: string;
  button: string;
  buttonAlt: string;
  text: string;
  textMuted: string;
  border: string;
  icon: string;
  progress: string;
  glass: string;
  logo: string;
}